fix migration

+++ db/migrate/20181016104953_add_inventory_amount_to_ingredients.rb
@@ -1,5 +1,5 @@
 class AddInventoryAmountToIngredients < ActiveRecord::Migration[5.2]
   def change
-    add_column :ingredients, :inventory_amount, :float(2)
+    add_column :ingredients, :inventory_amount, :float
   end
 end
